Create a data frame with stored digests and digests of current files

get_current_pdf_digests returns a data frame with a row for every file and columns for stored and current digests of source and output files.

Usage

get_current_pdf_digests(
  files,
  root_dir = NULL,
  static_path = "static",
  content_path = "content"
)

Arguments

files

A character vector of paths to source files (e.g., .Rmd).

Value

A a data frame with a row for every file and columns:

file

The source file name.

dest

The output file name.

alg

The digest algorithm.

digest

The stored digest for the source file.

dest_digest

The stored digest for the output file.

cur_digest

The digest for the current source file.

cur_dest_digest

The digest for the current output file.

Digests for missing files are set to NA.

Details

This function accepts a vector of source files and returns a data frame with a row for each file and columns for the stored digests and the digests of current source and output files.

See also