Determine the type of calendar entry from its calendar id.

Determine the type of calendar entry from its calendar id.

Usage

item_type(cal_id)

Arguments

cal_id

an integer calendar ID number.

Value

A string identifying the type of calendar entry. Current values are "class", "reading", "homework", "lab", "exam", "due date", "holiday", and "event".